+++ superset/charts/filters.py
@@ -26,10 +26,11 @@ from superset.connectors.sqla import models
 from superset.connectors.sqla.models import SqlaTable
 from superset.models.core import FavStar
 from superset.models.slice import Slice
+from superset.tags.filters import BaseTagIdFilter, BaseTagNameFilter
 from superset.utils.core import get_user_id
 from superset.utils.filters import get_dataset_access_filters
 from superset.views.base import BaseFilter
-from superset.views.base_api import BaseFavoriteFilter, BaseTagFilter
+from superset.views.base_api import BaseFavoriteFilter
 
 
 class ChartAllTextFilter(BaseFilter):  # pylint: disable=too-few-public-methods
@@ -60,7 +61,7 @@ class ChartFavoriteFilter(BaseFavoriteFilter):  # pylint: disable=too-few-public
     model = Slice
 
 
-class ChartTagNameFilter(BaseTagFilter):  # pylint: disable=too-few-public-methods
+class ChartTagNameFilter(BaseTagNameFilter):  # pylint: disable=too-few-public-methods
     """
     Custom filter for the GET list that filters all charts associated with
     a certain tag (by its name).
@@ -71,7 +72,7 @@ class ChartTagNameFilter(BaseTagFilter):  # pylint: disable=too-few-public-metho
     model = Slice
 
 
-class ChartTagIdFilter(BaseTagFilter):  # pylint: disable=too-few-public-methods
+class ChartTagIdFilter(BaseTagIdFilter):  # pylint: disable=too-few-public-methods
     """
     Custom filter for the GET list that filters all charts associated with
     a certain tag (by its ID).
@@ -79,7 +80,6 @@ class ChartTagIdFilter(BaseTagFilter):  # pylint: disable=too-few-public-methods
 
     arg_name = "chart_tag_id"
     class_name = "slice"
-    id_based_filter = True
     model = Slice
