+++ superset/views/base_api.py
@@ -31,7 +31,6 @@ from marshmallow import fields, Schema
 from sqlalchemy import and_, distinct, func
 from sqlalchemy.orm.query import Query
 
-from superset.connectors.sqla.models import SqlaTable
 from superset.exceptions import InvalidPayloadFormatError
 from superset.extensions import db, event_logger, security_manager, stats_logger_manager
 from superset.models.core import FavStar
@@ -40,7 +39,6 @@ from superset.models.slice import Slice
 from superset.schemas import error_payload_content
 from superset.sql_lab import Query as SqllabQuery
 from superset.superset_typing import FlaskResponse
-from superset.tags.models import Tag
 from superset.utils.core import get_user_id, time_function
 from superset.views.base import handle_api_exception
 
@@ -166,40 +164,6 @@ class BaseFavoriteFilter(BaseFilter):  # pylint: disable=too-few-public-methods
         if value:
             return query.filter(and_(self.model.id.in_(users_favorite_query)))
         return query.filter(and_(~self.model.id.in_(users_favorite_query)))
-
-
-class BaseTagFilter(BaseFilter):  # pylint: disable=too-few-public-methods
-    """
-    Base Custom filter for the GET list that filters all dashboards, slices
-    that a user has favored or not
-    """
-
-    name = _("Is tagged")
-    arg_name = ""
-    class_name = ""
-    """ The Tag class_name to user """
-    model: type[Dashboard | Slice | SqllabQuery | SqlaTable] = Dashboard
-    """ The SQLAlchemy model """
-    id_based_filter = False
-
-    def apply(self, query: Query, value: Any) -> Query:
-        # ID based filter
-        if self.id_based_filter:
-            tags_query = (
-                db.session.query(self.model.id)
-                .join(self.model.tags)
-                .filter(Tag.id == value)
-            )
-            return query.filter(self.model.id.in_(tags_query))
-
-        # Name based filter
-        ilike_value = f"%{value}%"
-        tags_query = (
-            db.session.query(self.model.id)
-            .join(self.model.tags)
-            .filter(Tag.name.ilike(ilike_value))
-        )
-        return query.filter(self.model.id.in_(tags_query))
 
 
 class BaseSupersetApiMixin:
